+++ DLSFTPClient/DemoApp/FileOperationsViewController.m
@@ -170,7 +170,7 @@ typedef enum {
 
     DLSFTPClientFailureBlock failureBlock = ^(NSError *error) {
         dispatch_async(dispatch_get_main_queue(), ^{
-            NSString *title = [NSString stringWithFormat:@"%@ Error: %d", error.domain, error.code];
+            NSString *title = [NSString stringWithFormat:@"%@ Error: %ld", error.domain, (long)error.code];
             UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:title
                                                                 message:[error localizedDescription]
                                                                delegate:nil
@@ -198,7 +198,7 @@ typedef enum {
 
     DLSFTPClientFailureBlock failureBlock = ^(NSError *error) {
         dispatch_async(dispatch_get_main_queue(), ^{
-            NSString *title = [NSString stringWithFormat:@"%@ Error: %d", error.domain, error.code];
+            NSString *title = [NSString stringWithFormat:@"%@ Error: %ld", error.domain, (long)error.code];
             UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:title
                                                                 message:[error localizedDescription]
                                                                delegate:nil
